feat(config): support configuration via pyproject.toml
Refactors coveralls/configuration.py into a package (ci, environment,
files, helpers) with __init__ as the entrypoint.

changelog:
coveralls-python now reads configuration from a ``[tool.coveralls]`` table
in ``pyproject.toml`` (no extra dependency required). The legacy
``.coveralls.yml`` is still supported and takes precedence when both exist.
