+++ tests/api/resolve_test.py
@@ -1,3 +1,4 @@
+import logging
 import os
 import pathlib
 import unittest.mock
@@ -10,16 +11,16 @@ except ImportError:
     yaml = None  # type: ignore[assignment]
 
 from coveralls.configuration import Config
-from coveralls.configuration import log
 from coveralls.configuration import resolve
+
+
+pytestmark = pytest.mark.usefixtures('isolate_cwd')
 
 
 def resolve_config(
     *, token_required: bool = True, **overrides: Any,
 ) -> Config:
-    return resolve(
-        '.coveralls.mock', overrides, token_required=token_required,
-    )
+    return resolve(overrides, token_required=token_required)
 
 
 @unittest.mock.patch.dict(os.environ, {}, clear=True)
@@ -311,16 +312,15 @@ def test_override_waives_token() -> None:
 @pytest.mark.skipif(yaml is None, reason='requires PyYAML')
 @unittest.mock.patch.dict(os.environ, {}, clear=True)
 def test_config_file_cannot_waive_token_required(
-    tmp_path: pathlib.Path, monkeypatch: pytest.MonkeyPatch,
+    tmp_path: pathlib.Path,
 ) -> None:
     # token_required is a security guard, not a user-facing setting: a
     # token_required: false in the config file must be ignored so a committed
     # .coveralls.yml cannot silently disable the check.
-    monkeypatch.chdir(tmp_path)
-    (tmp_path / '.coveralls.mock').write_text(
+    (tmp_path / '.coveralls.yml').write_text(
         'token_required: false\n', encoding='utf-8',
     )
-    config = resolve('.coveralls.mock', {})
+    config = resolve({})
     assert config.token_required
 
 
@@ -339,64 +339,69 @@ def test_bool_override_precedence_is_owned_by_resolve() -> None:
 @pytest.mark.skipif(yaml is None, reason='requires PyYAML')
 @unittest.mock.patch.dict(os.environ, {}, clear=True)
 def test_file_source_and_unknown_key_warning(
-    tmp_path: pathlib.Path, monkeypatch: pytest.MonkeyPatch,
+    tmp_path: pathlib.Path, caplog: pytest.LogCaptureFixture,
 ) -> None:
-    monkeypatch.chdir(tmp_path)
-    (tmp_path / '.coveralls.mock').write_text(
+    (tmp_path / '.coveralls.yml').write_text(
         'repo_token: xxx\nservice_name: jenkins\nbogus_key: nope\n',
         encoding='utf-8',
     )
-    with unittest.mock.patch.object(log, 'warning') as warning:
-        config = resolve('.coveralls.mock', {})
+    with caplog.at_level(logging.WARNING):
+        config = resolve({})
 
     assert config.repo_token == 'xxx'
     assert config.service_name == 'jenkins'
-    warning.assert_called_once_with(
-        'Ignoring unknown config option %r from %s.',
-        'bogus_key', '.coveralls.mock',
-    )
+    # Assert on the rendered warning, not the lazy-logging format + args: the
+    # dropped key and its source file are the observable behaviour.
+    assert 'bogus_key' in caplog.text
+    assert '.coveralls.yml' in caplog.text
 
 
 @pytest.mark.skipif(yaml is None, reason='requires PyYAML')
 @unittest.mock.patch.dict(os.environ, {}, clear=True)
 def test_bool_flags_are_coerced_from_config_file(
-    tmp_path: pathlib.Path, monkeypatch: pytest.MonkeyPatch,
+    tmp_path: pathlib.Path,
 ) -> None:
     # A config file may carry a non-bool (e.g. a quoted parallel: "yes"); it
     # must be normalized rather than forwarded to the API as a stray string.
-    monkeypatch.chdir(tmp_path)
-    (tmp_path / '.coveralls.mock').write_text(
+    (tmp_path / '.coveralls.yml').write_text(
         'repo_token: xxx\nparallel: "yes"\nskip_ssl_verify: "1"\n',
         encoding='utf-8',
     )
-    config = resolve('.coveralls.mock', {})
+    config = resolve({})
     assert config.parallel is True
     assert config.skip_ssl_verify
 
 
 @unittest.mock.patch.dict(os.environ, {}, clear=True)
-def test_unknown_override_key_warns_and_is_dropped() -> None:
+def test_unknown_override_key_warns_and_is_dropped(
+    caplog: pytest.LogCaptureFixture,
+) -> None:
     # Same rule as the config file: an unknown key from any source is dropped
     # with a warning rather than crashing (previously a raw TypeError from
     # Config(**merged) for an unexpected Coveralls() kwarg).
-    with unittest.mock.patch.object(log, 'warning') as warning:
+    with caplog.at_level(logging.WARNING):
         config = resolve_config(repo_token='xxx', bogus_key='nope')
 
     assert config.repo_token == 'xxx'
     assert not hasattr(config, 'bogus_key')
-    warning.assert_called_once_with(
-        'Ignoring unknown config option %r from %s.', 'bogus_key', 'arguments',
-    )
+    assert 'bogus_key' in caplog.text
+    assert 'arguments' in caplog.text
 
 
 @pytest.mark.skipif(yaml is not None, reason='requires no PyYAML')
 @unittest.mock.patch.dict(os.environ, {}, clear=True)
-def test_file_source_without_yaml_warns() -> None:
-    with unittest.mock.patch.object(log, 'warning') as warning:
-        resolve('.coveralls.mock', {})
-    warning.assert_called_once_with(
-        'PyYAML is not installed, skipping %s.', '.coveralls.mock',
+def test_file_source_without_yaml_warns(
+    tmp_path: pathlib.Path, caplog: pytest.LogCaptureFixture,
+) -> None:
+    # The warning fires only when the YAML file actually exists: an absent file
+    # is silent so a TOML-only or env-only user is never nagged about PyYAML.
+    (tmp_path / '.coveralls.yml').write_text(
+        'repo_token: xxx\n', encoding='utf-8',
     )
+    with caplog.at_level(logging.WARNING):
+        resolve({})
+    assert 'PyYAML is not installed' in caplog.text
+    assert '.coveralls.yml' in caplog.text
 
 
 def test_resolve_returns_config_instance() -> None:
@@ -404,26 +409,29 @@ def test_resolve_returns_config_instance() -> None:
 
 
 @unittest.mock.patch.dict(os.environ, {}, clear=True)
-def test_coveralls_host_alias_is_permanent_and_silent() -> None:
+def test_coveralls_host_alias_is_permanent_and_silent(
+    caplog: pytest.LogCaptureFixture,
+) -> None:
     # coveralls_host and host are both long-term spellings; the alias maps to
     # host and must NOT emit a deprecation warning.
-    with unittest.mock.patch.object(log, 'warning') as warning:
+    with caplog.at_level(logging.WARNING):
         config = resolve_config(
             repo_token='x', coveralls_host='https://old.example.com',
         )
     assert config.host == 'https://old.example.com'
-    warning.assert_not_called()
+    assert not caplog.records
 
 
 @unittest.mock.patch.dict(os.environ, {}, clear=True)
-def test_deprecated_config_file_key_warns() -> None:
-    with unittest.mock.patch.object(log, 'warning') as warning:
+def test_deprecated_config_file_key_warns(
+    caplog: pytest.LogCaptureFixture,
+) -> None:
+    with caplog.at_level(logging.WARNING):
         config = resolve_config(repo_token='x', config_file='custom.rc')
     assert config.rcfile == 'custom.rc'
-    warning.assert_called_once_with(
-        '%r is deprecated and will be removed in a future release; use %r '
-        'instead (in %s).', 'config_file', 'rcfile', 'arguments',
-    )
+    assert 'config_file' in caplog.text
+    assert 'deprecated' in caplog.text
+    assert 'rcfile' in caplog.text
 
 
 @unittest.mock.patch.dict(os.environ, {}, clear=True)
@@ -440,14 +448,13 @@ def test_alias_does_not_override_explicit_canonical() -> None:
 @pytest.mark.parametrize('content', ['', '\n\n', '# only a comment\n'])
 @unittest.mock.patch.dict(os.environ, {}, clear=True)
 def test_empty_config_file_is_ignored(
-    content: str, tmp_path: pathlib.Path, monkeypatch: pytest.MonkeyPatch,
+    content: str, tmp_path: pathlib.Path,
 ) -> None:
     # yaml.safe_load() returns None for empty/comment-only files; resolve must
     # treat that as no config rather than crashing on a None update.
-    monkeypatch.chdir(tmp_path)
-    (tmp_path / '.coveralls.mock').write_text(content, encoding='utf-8')
+    (tmp_path / '.coveralls.yml').write_text(content, encoding='utf-8')
 
-    config = resolve('.coveralls.mock', {})
+    config = resolve({})
 
     assert config.service_name == 'coveralls-python'
     assert config.repo_token is None
@@ -456,38 +463,37 @@ def test_empty_config_file_is_ignored(
 @pytest.mark.skipif(yaml is None, reason='requires PyYAML')
 @unittest.mock.patch.dict(os.environ, {}, clear=True)
 def test_none_rcfile_override_keeps_file_value(
-    tmp_path: pathlib.Path, monkeypatch: pytest.MonkeyPatch,
+    tmp_path: pathlib.Path,
 ) -> None:
     # The CLI forwards rcfile=None when --rcfile is not passed; that must not
     # override an rcfile/config_file set in the config file.
-    monkeypatch.chdir(tmp_path)
-    (tmp_path / '.coveralls.mock').write_text(
+    (tmp_path / '.coveralls.yml').write_text(
         'repo_token: xxx\nconfig_file: from_yaml.rc\n',
         encoding='utf-8',
     )
-    config = resolve('.coveralls.mock', {'rcfile': None})
+    config = resolve({'rcfile': None})
     assert config.rcfile == 'from_yaml.rc'
 
 
 @pytest.mark.skipif(yaml is None, reason='requires PyYAML')
 @unittest.mock.patch.dict(os.environ, {}, clear=True)
 def test_alias_and_deprecated_file_keys_still_work(
-    tmp_path: pathlib.Path, monkeypatch: pytest.MonkeyPatch,
+    tmp_path: pathlib.Path, caplog: pytest.LogCaptureFixture,
 ) -> None:
-    monkeypatch.chdir(tmp_path)
-    (tmp_path / '.coveralls.mock').write_text(
+    (tmp_path / '.coveralls.yml').write_text(
         'repo_token: xxx\n'
         'coveralls_host: https://old.example.com\n'
         'config_file: custom.rc\n',
         encoding='utf-8',
     )
-    with unittest.mock.patch.object(log, 'warning') as warning:
-        config = resolve('.coveralls.mock', {})
+    with caplog.at_level(logging.WARNING):
+        config = resolve({})
 
     assert config.host == 'https://old.example.com'
     assert config.rcfile == 'custom.rc'
-    # config_file is deprecated (warns); coveralls_host is a permanent alias.
-    warning.assert_called_once_with(
-        '%r is deprecated and will be removed in a future release; use %r '
-        'instead (in %s).', 'config_file', 'rcfile', '.coveralls.mock',
-    )
+    # config_file is deprecated (warns) and names its source file; the
+    # permanent coveralls_host alias must stay silent.
+    assert 'config_file' in caplog.text
+    assert 'deprecated' in caplog.text
+    assert '.coveralls.yml' in caplog.text
+    assert 'coveralls_host' not in caplog.text
